// packages/job-scheduling/src/utils/types.ts
export interface ICatalystJobMetaConfig {
	/** Number of time the job can be retied incase of failure */
	number_of_retries: number;
	/** Interval in which the job should be retried */
	retry_interval: number;
}
// ...
export interface ICatalystJobBasic extends ICatalystJSON {
	/** Name of the job */
	job_name: string;
	/** Configuration of the job */
	job_config?: ICatalystJobMetaConfig;
	/** Id of the jobpool to which the job should be submitted */
	jobpool_id?: string;
	/** Name of the jobpool to which the job should be submitted */
	jobpool_name?: string;
	/** Source type of the job */
	source_type?: JOB_SOURCE_TYPE;
}

/**
 * Job that's associated with a job function as a target and executed in a function jobpool
 */
export interface ICatalystFunctionJob extends ICatalystJobBasic {
	/** Target type for function jobs */
	target_type: TARGET_TYPE.FUNCTION;
	/** Target function's Id */
	target_id?: string;
	/** Target function's name */
	target_name?: string;
	/** Params that should be passed to the function */
	params?: Record<string, string>;
}